Add documentation and format JavaDocs

+++ src/main/java/com/swiftelan/tag/paging/PagingTag.java
@@ -18,10 +18,11 @@ import com.swiftelan.tag.util.Strings;
 /**
  * Tag handler for rendering a paging control.
  * <p>
- * A component that uses a list of links for paging. A range of pages around the current page is rendered. By default a first and
- * last page links are also displayed. Next and previous page links can be included by using the {@link NextTag next} and
- * {@link PreviousTag previous} tags in the body.
+ * A component that uses a list of links for paging. A range of pages around the current page is rendered. By default a
+ * first and last page links are also displayed. Next and previous page links can be included by using the
+ * {@link NextTag next} and {@link PreviousTag previous} tags in the body.
  * </p>
+ *
  * @see NextTag
  * @see PreviousTag
  */
@@ -41,9 +42,9 @@ public class PagingTag extends ComponentTagSupport implements JspIdConsumer {
 
 	@Override
 	public void doTag() throws JspException, IOException {
-		totalNumberOfItems = (Integer) _totalNumberOfItems.getValue(getJspContext().getELContext());
-		pageSize = (Integer) _pageSize.getValue(getJspContext().getELContext());
-		firstResult = (Integer) _firstResult.getValue(getJspContext().getELContext());
+		totalNumberOfItems = ((Integer) _totalNumberOfItems.getValue(getJspContext().getELContext())).intValue();
+		pageSize = ((Integer) _pageSize.getValue(getJspContext().getELContext())).intValue();
+		firstResult = ((Integer) _firstResult.getValue(getJspContext().getELContext())).intValue();
 		if (pageSize == 0) {
 			currentPage = 1;
 		} else {
@@ -141,17 +142,21 @@ public class PagingTag extends ComponentTagSupport implements JspIdConsumer {
 		String uri;
 		String query;
 		switch (getRequest().getDispatcherType()) {
-			case FORWARD:
-				uri = (String) getJspContext().getAttribute(RequestDispatcher.FORWARD_REQUEST_URI, PageContext.REQUEST_SCOPE);
-				query = (String) getJspContext().getAttribute(RequestDispatcher.FORWARD_QUERY_STRING, PageContext.REQUEST_SCOPE);
-				break;
-			case INCLUDE:
-				uri = (String) getJspContext().getAttribute(RequestDispatcher.INCLUDE_REQUEST_URI, PageContext.REQUEST_SCOPE);
-				query = (String) getJspContext().getAttribute(RequestDispatcher.INCLUDE_QUERY_STRING, PageContext.REQUEST_SCOPE);
-				break;
-			default:
-				uri = getRequest().getRequestURI();
-				query = getRequest().getQueryString();
+		case FORWARD:
+			uri = (String) getJspContext().getAttribute(RequestDispatcher.FORWARD_REQUEST_URI,
+					PageContext.REQUEST_SCOPE);
+			query = (String) getJspContext().getAttribute(RequestDispatcher.FORWARD_QUERY_STRING,
+					PageContext.REQUEST_SCOPE);
+			break;
+		case INCLUDE:
+			uri = (String) getJspContext().getAttribute(RequestDispatcher.INCLUDE_REQUEST_URI,
+					PageContext.REQUEST_SCOPE);
+			query = (String) getJspContext().getAttribute(RequestDispatcher.INCLUDE_QUERY_STRING,
+					PageContext.REQUEST_SCOPE);
+			break;
+		default:
+			uri = getRequest().getRequestURI();
+			query = getRequest().getQueryString();
 		}
 
 		StringBuilder sb = new StringBuilder(uri);
@@ -211,6 +216,7 @@ public class PagingTag extends ComponentTagSupport implements JspIdConsumer {
 
 	/**
 	 * Set the expression that will be used to get the number of items displayed on a single page.
+	 *
 	 * @param pageSize Expression for the page size
 	 */
 	public void setPageSize(ValueExpression pageSize) {
@@ -218,7 +224,9 @@ public class PagingTag extends ComponentTagSupport implements JspIdConsumer {
 	}
 
 	/**
-	 * Set the expression that will be used to get the total number of items in the result set if paging were not being used.
+	 * Set the expression that will be used to get the total number of items in the result set if paging were not being
+	 * used.
+	 *
 	 * @param totalNumberOfItems Expression for the total number of items
 	 */
 	public void setTotalNumberOfItems(ValueExpression totalNumberOfItems) {
@@ -227,6 +235,7 @@ public class PagingTag extends ComponentTagSupport implements JspIdConsumer {
 
 	/**
 	 * Set the expression that will be used to get the index of the first item displayed in the current page.
+	 *
 	 * @param firstResult Expression for the index of the first item
 	 */
 	public void setFirstResult(ValueExpression firstResult) {
